+++ src/clj_kondo/impl/linters.clj
@@ -634,6 +634,34 @@
                         :type :unused-value
                         :message "Unused value"}))))))))))))
 
+(defn- lint-aliased-referred-var! [ctx ns]
+  (when-not (utils/linter-disabled? ctx :aliased-referred-var)
+    (let [{:keys [referred-vars]} ns]
+      (when (seq referred-vars)
+        (let [referred-by-full-name (->> referred-vars
+                                       (map (fn [[k info]]
+                                              [[(str (:ns info))
+                                                (str (:name info))]
+                                               k]))
+                                       (into {}))]
+          (doseq [{:keys [alias resolved-ns] :as usage} (:used-vars ns)
+                  :let [v-name (str (:name usage))
+                        v-ns (str (or resolved-ns (:to usage)))
+                        full-name [v-ns v-name]
+                        referred-name (when alias
+                                        (get referred-by-full-name full-name))]
+                  :when referred-name
+                  :let [msg (format "var %s/%s is referred as %s but also used via alias %s"
+                                    (first full-name) (second full-name)
+                                    referred-name alias)]]
+            (findings/reg-finding!
+             ctx
+             {:filename (:filename ns)
+              :row (:row usage)
+              :col (:col usage)
+              :type :aliased-referred-var
+              :message msg})))))))
+
 (defn lint-unused-namespaces!
   [ctx idacs]
   (let [config (:config ctx)]
@@ -722,7 +750,8 @@
           (findings/reg-finding!
            ctx
            (node->line (:filename (meta alias)) alias
-                       :unused-alias (str "Unused alias: " alias))))))))
+                       :unused-alias (str "Unused alias: " alias)))))
+      (lint-aliased-referred-var! ctx ns))))
 
 (defn lint-unused-excluded-vars! [ctx]
   (when-not (linter-disabled? ctx :unused-excluded-var)